refactor: move validate functions into utils
