+++ src/constants/countries.ts
@@ -1,0 +1,1 @@
+export const COUNTRIES = ['United Kingdom', 'Germany', 'United States'];
