+++ src/features/profile/components/addresses/address-edit-modal/edit-address-modal.tsx
@@ -5,6 +5,7 @@ import { Button, Checkbox, Group, LoadingOverlay, Modal, SimpleGrid } from '@man
 import { BaseButton } from '@/components/base-button';
 import { CustomSelect } from '@/components/custom-select';
 import { CustomTextInput } from '@/components/custom-text-input';
+import { COUNTRIES } from '@/constants/countries';
 import { AddressEditModalProps } from '@/features/profile/types/edit-modal-props';
 import { UserAddress } from '@/features/profile/types/user-address';
 import { addNotification } from '@/utils/show-notification';
@@ -18,8 +19,6 @@ import {
 } from '../../../api/address-api';
 
 import classes from '@/components/modals/modal.module.css';
-
-const COUNTRIES = ['United Kingdom', 'Germany', 'United States'];
 
 const EditAddressModal = ({
   addresses,
