+++ src/features/registration-page/registration-page.tsx
@@ -4,20 +4,26 @@ import { Link, useNavigate } from 'react-router-dom';
 
 import { Address } from '@commercetools/platform-sdk';
 import { Anchor, Checkbox, Container, LoadingOverlay, SimpleGrid, Text, Title } from '@mantine/core';
-import { UseFormReturnType, useForm } from '@mantine/form';
+import { useForm } from '@mantine/form';
 import { useDisclosure } from '@mantine/hooks';
 import dayjs from 'dayjs';
-import { postcodeValidator } from 'postcode-validator';
 
 import { BaseButton } from '@/components/base-button';
 import { CustomDateInput } from '@/components/custom-date-input';
 import { CustomPasswordInput } from '@/components/custom-password-input';
 import { CustomSelect } from '@/components/custom-select';
 import { CustomTextInput } from '@/components/custom-text-input';
+import { COUNTRIES } from '@/constants/countries';
 import { createCustomer } from '@/lib/commerstools/customer-creator';
 import { AuthState } from '@/types/authState';
+import { validatePassword } from '@/utils';
 import { addNotification } from '@/utils/show-notification';
-import { validatePassword } from '@/utils/validate-password';
+import { isProperCountry } from '@/utils/validate/is-proper-country';
+import { isProperPostcode } from '@/utils/validate/is-proper-postcode';
+import { isEmail } from '@/utils/validate/isEmail';
+import { notEmpty } from '@/utils/validate/not-empty';
+import { onlyLetters } from '@/utils/validate/only-letters';
+import { transformCountryIntoCountryCode } from '@/utils/validate/transform-country';
 
 import { setAuthState } from '../auth/authSlice';
 import { postCustomerLogin } from '../login-page/api';
@@ -36,53 +42,14 @@ interface CheckboxProps {
   value?: string;
 }
 
-const COUNTRIES = ['United Kingdom', 'Germany', 'United States'];
 const TODAY = new Date();
 const ONE_HUNDRED_AND_THIRTY_YEARS_AGO = dayjs(new Date()).subtract(130, 'year').toDate();
-const notEmpty = (value: string): null | string => (value.trim() ? null : 'Required field');
-
-const onlyLetters =
-  (message: string) =>
-  (value: string): null | string => {
-    if (!value) {
-      return 'Required field';
-    }
-    if (!/^[A-Za-zäöüßÄÖÜА-Яа-я]+$/.test(value)) {
-      return message;
-    }
-    return null;
-  };
-
-const isEmail =
-  (message: string) =>
-  (value: string): null | string => {
-    if (!value) {
-      return 'Required field';
-    }
-
-    return /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/.test(value) ? null : message;
-  };
 
 const matchesPassword = (value: string, values: { password: string }): null | string => {
   if (!value) {
     return 'Required field';
   }
   return value !== values.password ? 'Passwords did not match' : null;
-};
-
-const isProperCountry = (value: string): null | string => (COUNTRIES.includes(value) ? null : 'Invalid country');
-
-const transformCountryIntoCountryCode = (country: string): string => {
-  switch (country) {
-    case 'Germany':
-      return 'DE';
-    case 'United Kingdom':
-      return 'UK';
-    case 'United States':
-      return 'US';
-    default:
-      return '';
-  }
 };
 
 const wrapSameAddressCheck =
@@ -92,16 +59,6 @@ const wrapSameAddressCheck =
       return null;
     }
     return cb(value, values);
-  };
-
-const isProperPostcode =
-  <K extends string, T extends Record<K, string>>(countryField: K) =>
-  (value: string, values: UseFormReturnType<T>['values']): null | string => {
-    const code = transformCountryIntoCountryCode(values[countryField]);
-    if (!code) {
-      return 'Invalid country';
-    }
-    return postcodeValidator(value, code) ? null : 'Invalid postcode';
   };
 
 const RegistrationPage: FC = () => {
