+++ src/utils/validate/is-proper-country.ts
@@ -1,0 +1,3 @@
+import { COUNTRIES } from '@/constants/countries';
+
+export const isProperCountry = (value: string): null | string => (COUNTRIES.includes(value) ? null : 'Invalid country');
