+++ src/utils/validate/not-empty.ts
@@ -1,0 +1,1 @@
+export const notEmpty = (value: string): null | string => (value.trim() ? null : 'Required field');
