+++ src/utils/validate/only-letters.ts
@@ -1,0 +1,11 @@
+export const onlyLetters =
+  (message: string) =>
+  (value: string): null | string => {
+    if (!value) {
+      return 'Required field';
+    }
+    if (!/^[A-Za-zäöüßÄÖÜА-Яа-я]+$/.test(value)) {
+      return message;
+    }
+    return null;
+  };
